feat: impl some io function
